appling caAERS CAS patch ..Following are comments fron Joshua
"The problem that caAERS is having with WebSSO is that the CAS client that caAERS (and C3PR and PSC) is using can't validate the hostname that is specified in the host certificate of the WebSSO server. If caAERS were using the CAS client that is provided by WebSSO, we'd be able to just make a configuration change, but since caAERS is using the client that is packaged with Acegi 1.0.3, we need to provide a different implementation of TicketValidator which provides an instance of HostnameVerify to verify Globus-style hostnames".


git-svn-id: https://ncisvn.nci.nih.gov/svn/psc/trunk@2955 0d517254-b314-0410-acde-c619094fa49f