Part of the commit for # 683 - The following modifications are added:
a) made population part of the accordion be able to display subject on different studies.
b) used a subjectAssignedIdentifier to distinguish what subject belongs to, since subject can belong to the same study multiple times.
c) used a subjectAssignedIdentifier in markActivity part as well, in a drop down study selector

git-svn-id: https://ncisvn.nci.nih.gov/svn/psc/trunk@4011 0d517254-b314-0410-acde-c619094fa49f