Move first-pass authorization from AuthorizingFinder to an overridden #handle method on AbstractPscResource.  #743.
This was necessary because in Restlet 2, the #find method in Finder is effectively prevented from changing the response status or stopping execution of the Restlet chain.

This change required corrections in several tests which were testing the wrong things (e.g., expecting 404 when 403 should have actually be returned).  A benefit of this change is that resource unit tests can now test both the first pass authorization and any detailed authorization performed in the resource itself.

git-svn-id: https://ncisvn.nci.nih.gov/svn/psc/trunk@5967 0d517254-b314-0410-acde-c619094fa49f