The docs resource is available from two different base URIs -- docs and docs/psc.  #743.
In Restlet 1.1 this was handled by the default router match mode (STARTS_WITH).  In Restlet 2.0, the default match mode is EQUALS.  This is a good change, so I just manually mapped the docs resource to both paths.

This commit also includes some cleanup of the unit test for this resource.

git-svn-id: https://ncisvn.nci.nih.gov/svn/psc/trunk@5968 0d517254-b314-0410-acde-c619094fa49f