Begin improving template difference messages. #1444.
This commit introduces a new interface (DeepComparable) for things that
implement deepEquals, but does not implement it everywhere. Similarly, it
starts rewording error messages, but does not fix them at all levels.

Finally, it introduces a tree representation of the differences list as a
stopgap to make them slightly more legible and uses it in TemplateImportService.

git-svn-id: https://ncisvn.nci.nih.gov/svn/psc/branches/releases/2.9-stable@6812 0d517254-b314-0410-acde-c619094fa49f